Move out of __init__.py to cli.py
